FizzBuzz confusion
I tried several different code combinations to make this work and even with the flowchart, I couldn't figure out the right approach. The closest I was able to get was getting "Fizz" to print several times in a row. I'll go back over the material and seek out a tutor if needed.